refactor(e2e) global config 'isAngular' in package.json
By default, Protractor waits for angular to finish its work.
When you use it without angular, you need to disable the browser.synchronisation.
This config is here to do that job globally (and not have to redo it),
if you have a full angular or full non-angular website.
If you have a mixed site, a beforEachIsAngular helper is available in `test/e2e/utils.js`.