fix(test) timeout on click using MutationObserver (temporary)
This is a temp commit.
In order to avoid using setTimeouts in unit-tests,
I tried using DOM3 MutationObserver.
I made a onDOMElementChange utility in `test/unit/spec`

But since the shim for MutationObserver messes with jasmine async business, I'll remove it on the next commit
https://github.com/megawac/MutationObserver.js

So you can launch the tests with ./node_modules/.bin/karma start --browsers Chrome@@